+++ src/src/Controller/ApiBookController.php
@@ -50,19 +50,24 @@ class ApiBookController extends BaseController
     {
         $data = $this->getRequestJSON();
         $response = new JsonResponse();
-        $bookService->setOperation('create');
-        //Валидация 
-        $validate = $bookService->Validate($response, $data);
-        if ($validate['res']) {
-            //Проверка на уникальность и запись в бд
-            $create = $bookService->Create($em, $response, $data);
-            if ($create['res']) {
-                $response->setStatusCode(200);
-                $response->setData(['message' => 'Success']);
+        if (array_key_exists('name', $data) and array_key_exists('author', $data)) {
+            $bookService->setOperation('create');
+            //Валидация 
+            $validate = $bookService->Validate($response, $data);
+            if ($validate['res']) {
+                //Проверка на уникальность и запись в бд
+                $create = $bookService->Create($em, $response, $data);
+                if ($create['res']) {
+                    $response->setStatusCode(200);
+                    $response->setData(['message' => 'Success']);
+                }
+                return $create['response'];
+            } else {
+                return $validate['response'];
             }
-            return $create['response'];
         } else {
-            return $validate['response'];
+            $response->setStatusCode(422);
+            $response->setData(['message' => 'Not enough parameters']);
         }
         return $response;
     }
@@ -103,19 +108,24 @@ class ApiBookController extends BaseController
     {
         $data = $this->getRequestJSON();
         $response = new JsonResponse();
-        //Валидация 
-        $bookService->setOperation('update');
-        $validate = $bookService->Validate($response, $data);
-        if ($validate['res']) {
-            //Проверка на уникальность и запись в бд
-            $update = $bookService->Update($em, $response, $data);
-            if ($update['res']) {
-                $response->setStatusCode(200);
-                $response->setData(['message' => 'Success']);
+        if (array_key_exists('id', $data) and array_key_exists('name', $data) and array_key_exists('author', $data)) {
+            //Валидация 
+            $bookService->setOperation('update');
+            $validate = $bookService->Validate($response, $data);
+            if ($validate['res']) {
+                //Проверка на уникальность и запись в бд
+                $update = $bookService->Update($em, $response, $data);
+                if ($update['res']) {
+                    $response->setStatusCode(200);
+                    $response->setData(['message' => 'Success']);
+                }
+                return $update['response'];
+            } else {
+                return $validate['response'];
             }
-            return $update['response'];
         } else {
-            return $validate['response'];
+            $response->setStatusCode(422);
+            $response->setData(['message' => 'Not enough parameters']);
         }
 
         return $response;
@@ -190,21 +200,25 @@ class ApiBookController extends BaseController
     {
         $response = new JsonResponse();
         $data = $this->getRequestJSON();
-        //Валидация 
-        $bookService->setOperation('delete');
-        $validate = $bookService->Validate($response, $data);
-        if ($validate['res']) {
-            //Проверка на уникальность и запись в бд
-            $delete = $bookService->Delete($em, $response, $data);
-            if ($delete['res']) {
-                $response->setStatusCode(200);
-                $response->setData(['message' => 'Success']);
+        if (array_key_exists('id', $data)) {
+            //Валидация 
+            $bookService->setOperation('delete');
+            $validate = $bookService->Validate($response, $data);
+            if ($validate['res']) {
+                //Проверка на уникальность и запись в бд
+                $delete = $bookService->Delete($em, $response, $data);
+                if ($delete['res']) {
+                    $response->setStatusCode(200);
+                    $response->setData(['message' => 'Success']);
+                }
+                return $delete['response'];
+            } else {
+                return $validate['response'];
             }
-            return $delete['response'];
         } else {
-            return $validate['response'];
+            $response->setStatusCode(422);
+            $response->setData(['message' => 'Not enough parameters']);
         }
-
         return $response;
     }
     /**
